Added MaterialFX
[MaterialFX](https://github.com/palexdev/MaterialFX)
Born as an alternative for the popular JFoenix library which is not maintained anymore and too buggy especially for newer versions of Java, MaterialFX maintains the same base goal, making JavaFX's controls better both in appearance and in functionality, but also brings a lot of new features.
New utilities from StringUtils to JavaFX utils such as: NodeUtils, LabelUtils, DragResizer and so on.
Controls built from scratch: in my library there is the concept of legacy and "brand new". Legacy controls are controls which extend JavaFX's ones but with a modern design and even new functionalities. The "brand new" controls are controls which extends the base class Control, all the functionalities are re-defined from scratch, the advantage of this approach is that I have full control on the source code, so I can bring more easily new features and bug fixes and even improve them in performance.
Some of these controls are: list views based on the popular Flowless, new TreeViews (will be reworked probably for better performance), table views, even context menus, and so on.